+++ src/main/java/com/deltasf/createpropulsion/utility/tooltips/TooltipModifiers.java
@@ -22,13 +22,13 @@ public class TooltipModifiers {
     static {
         //Thruster
         tooltipModificationLookup.put(PropulsionBlocks.THRUSTER_BLOCK.asItem(), (payload) -> {
-            float thrustMultiplier = (float)(double)PropulsionConfig.THRUSTER_THRUST_MULTIPLIER.get();
+            float thrustMultiplier = PropulsionConfig.THRUSTER_THRUST_MULTIPLIER.get().floatValue();
             int thrusterStrength = Math.round(ThrusterBlockEntity.BASE_MAX_THRUST / 1000.0f * thrustMultiplier);
             return Component.translatable(payload.path + ".tooltip.summary").getString().replace("{}", String.valueOf(thrusterStrength));
         });
         //Creative thruster
         tooltipModificationLookup.put(PropulsionBlocks.CREATIVE_THRUSTER_BLOCK.asItem(), (payload) -> {
-            float thrustMultiplier = (float)(double)PropulsionConfig.CREATIVE_THRUSTER_THRUST_MULTIPLIER.get();
+            float thrustMultiplier = PropulsionConfig.CREATIVE_THRUSTER_THRUST_MULTIPLIER.get().floatValue();
             int thrusterStrength = Math.round(1000 * thrustMultiplier);
             return Component.translatable(payload.path + ".tooltip.summary").getString().replace("{}", String.valueOf(thrusterStrength));
         });
